# Bugfix recursive object index