--- a/parser.go
+++ b/parser.go
@@ -122,7 +122,6 @@ func searchKeys(data []byte, keys ...string) int {
 	}
 
 	var stackbuf [unescapeStackBufSize]byte // stack-allocated array for allocation-free unescaping of small strings
-
 	for i < ln {
 		switch data[i] {
 		case '"':
@@ -195,7 +194,11 @@ func searchKeys(data []byte, keys ...string) int {
 				if valueFound == nil {
 					return -1
 				} else {
-					return i + valueOffset + searchKeys(valueFound, keys[level+1:]...)
+					subIndex := searchKeys(valueFound, keys[level+1:]...)
+					if subIndex < 0 {
+						return -1
+					}
+					return i + valueOffset + subIndex
 				}
 			} else {
 				// Do not search for keys inside arrays