**Description**: What this PR does
Fix issue #81 
Previously offset from nested searchKeys call was added to parent value such that 11 + (-1) would return a "successful" value of 10. I added logic to read the value of the nested call and return a failure as needed.

**Benchmark before change**:
BenchmarkJsonParserLarge-8                        200000             44154 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-8                      1000000              7197 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-8         2000000              4168 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-8         2000000              4527 ns/op             656 B/op         13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-8      1000000              7091 ns/op             608 B/op         12 allocs/op
BenchmarkJsonParserSmall-8                      10000000               605 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-8         20000000               445 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-8         10000000               684 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-8      20000000               596 ns/op             240 B/op          8 allocs/op

**Benchmark after change**:
BenchmarkJsonParserLarge-8                        200000             44139 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-8                      1000000              7075 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-8         2000000              4161 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-8         2000000              4524 ns/op             656 B/op         13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-8      1000000              7056 ns/op             608 B/op         12 allocs/op
BenchmarkJsonParserSmall-8                      10000000               604 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-8         20000000               445 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-8         10000000               682 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-8      20000000               598 ns/op             240 B/op          8 allocs/op