@jensenak 2020-03-13
> Hello, how should I run the benchmark, at the moment I am running, I get an error. Do you know why?

I just ran `make bench` as directed in the readme. However, this PR was merged 3 years ago, so I don't remember what go version and what configuration options I might have been using.

The file that's throwing those errors has a note at the top: `THIS FILE IS AUTO-GENERATED BY codecgen.` and the version of the file committed to this repo is 4 years old. It calls exported functions like `codec.genHelperEncoder` that have an explicit note in source: `Library users: DO NOT USE IT DIRECTLY. IT WILL CHANGE CONTINOUSLY WITHOUT NOTICE.` (https://github.com/ugorji/go/blob/42bc974514ff101a54c6b72a0e4dee29d96c0b26/codec/gen-helper.generated.go#L22)

Though I didn't generate the benchmark_codecgen.go file, I'm guessing it needs to be regenerated with a more recent version of the codec library?